fix: cv2.line coordinates must be int
Cast image coordinates to int for drawing lines.
No idea why this was not an issue in the past (it is already done like
this in another cv2.line call a few lines below).

+++ scripts/calibrate_trifingerpro_cameras.py
@@ -258,8 +258,8 @@ def calibrate_mean_extrinsic_parameters(
             for p1, p2 in point_pairs:
                 cv2.line(
                     img,
-                    tuple(imgpoints[p1, 0]),
-                    tuple(imgpoints[p2, 0]),
+                    tuple(imgpoints[p1, 0].astype(int)),
+                    tuple(imgpoints[p2, 0].astype(int)),
                     [200, 200, 0],
                     thickness=2,
                 )
